requirements.txt
APScheduler==3.6.3
blinker==1.6.2
cachetools==4.2.2
certifi==2023.7.22
click==8.1.7
Flask==3.0.0
tinydb
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
tornado==6.1
tzlocal==5.0.1
Werkzeug==3.0.0
requests
reportlab